requirements.txt
django==4.1.3
django-ninja
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
distlib==0.3.5
filelock==3.8.0
idna==3.4
maturin==0.13.7
mnemonic==0.20
numpy==1.23.5
pandas==1.5.1
pipenv==2022.8.17
platformdirs==2.5.2
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.3
virtualenv-clone==0.5.7
django-admin-interface==0.22.2
tomlkit~=0.11.6
coincurve~=18.0.0
pycryptodome
psutil~=5.9.4
django-ipware
django-rq
humanfriendly~=10.0
passpy~=1.0.2
killport~=1.2.0
sentry-sdk
django-ninja-apikey
django-recaptcha
asgiref~=3.5.2
rq~=1.11.1
setuptools~=65.6.3
django-hitcount